Remove support for Python 3.6
Entailed a pyupgrade run which pretty much only resulted in a few more
f-strings.

Also remove all mentions of 3.6 in the docs, scripts etc.
This more or less was a "find and replace" operation. That means some
"example output" with python versions was also adapted. That output
should probably be "completely" updated, but for now, I thought it
shouldn't say anything about Python 3.6 when typer no longer supports
it.